refactor input reading
