+++ examples/hello-world.psl
@@ -1,1 +1,5 @@
-122 + 2 + 3 * 4 +
+{
+    5 * {
+5*5
+}
+}
