+++ proteus-compiler-kotlin/src/main/kotlin/lang/proteus/api/input/ProteusSourceTextInputReader.kt
@@ -1,0 +1,5 @@
+package lang.proteus.api.input
+
+abstract class ProteusSourceTextInputReader {
+    abstract fun read(): String?
+}
